Move api code to module
