+++ front/app/modules/commercial/project_description_builder/admin/components/ContentViewer/Viewer/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 // hooks
-import useProjectDescriptionBuilderLayout from 'api/project_description_builder/useProjectDescriptionBuilderLayout';
+import useProjectDescriptionBuilderLayout from 'modules/commercial/project_description_builder/api/useProjectDescriptionBuilderLayout';
 import useLocale from 'hooks/useLocale';
 import useLocalize from 'hooks/useLocalize';
 import useProjectFiles from 'api/project_files/useProjectFiles';
