+++ front/app/modules/commercial/project_description_builder/admin/components/ProjectDescriptionBuilderToggle/ProjectDescriptionBuilderToggle.test.tsx
@@ -15,7 +15,7 @@ const mockProjectDescriptionBuilderLayoutData: typeof DEFAULT_PROJECT_DESCRIPTIO
   DEFAULT_PROJECT_DESCRIPTION_BUILDER_LAYOUT_DATA;
 
 jest.mock(
-  'api/project_description_builder/useProjectDescriptionBuilderLayout',
+  'modules/commercial/project_description_builder/api/useProjectDescriptionBuilderLayout',
   () => () => {
     return {
       data: mockProjectDescriptionBuilderLayoutData,
@@ -25,7 +25,7 @@ jest.mock(
 
 const mockAddProjectDescriptionBuilderLayout = jest.fn();
 jest.mock(
-  'api/project_description_builder/useAddProjectDescriptionBuilderLayout',
+  'modules/commercial/project_description_builder/api/useAddProjectDescriptionBuilderLayout',
   () => jest.fn(() => ({ mutateAsync: mockAddProjectDescriptionBuilderLayout }))
 );
