+++ front/app/modules/commercial/project_description_builder/admin/components/ProjectDescriptionBuilderTopBar/index.tsx
@@ -4,7 +4,7 @@ import React, { useState } from 'react';
 import useProjectById from 'api/projects/useProjectById';
 import useLocalize from 'hooks/useLocalize';
 import { useEditor, SerializedNodes } from '@craftjs/core';
-import useAddProjectDescriptionBuilderLayout from 'api/project_description_builder/useAddProjectDescriptionBuilderLayout';
+import useAddProjectDescriptionBuilderLayout from 'modules/commercial/project_description_builder/api/useAddProjectDescriptionBuilderLayout';
 
 // components
 import Container from 'components/admin/ContentBuilder/TopBar/Container';
