+++ front/app/modules/commercial/project_description_builder/admin/containers/FullscreenPreview/FullscreenPreview.test.tsx
@@ -21,7 +21,7 @@ let mockProjectDescriptionBuilderLayoutData:
   | Error = DEFAULT_PROJECT_DESCRIPTION_BUILDER_LAYOUT_DATA;
 
 jest.mock(
-  'api/project_description_builder/useProjectDescriptionBuilderLayout',
+  'modules/commercial/project_description_builder/api/useProjectDescriptionBuilderLayout',
   () => () => {
     return {
       data: mockProjectDescriptionBuilderLayoutData,
