+++ front/app/modules/commercial/project_description_builder/admin/containers/FullscreenPreview/index.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 
 // hooks
-import useProjectDescriptionBuilderLayout from 'api/project_description_builder/useProjectDescriptionBuilderLayout';
+import useProjectDescriptionBuilderLayout from 'modules/commercial/project_description_builder/api/useProjectDescriptionBuilderLayout';
 import useLocale from 'hooks/useLocale';
 import useProjectById from 'api/projects/useProjectById';
 import { useParams } from 'react-router-dom';
