+++ front/app/modules/commercial/project_description_builder/api/__mocks__/projectDescriptionBuilderLayout.ts
