+++ front/app/modules/commercial/project_description_builder/api/types.ts
